统一主键为雪花ID，前后端全面切换为字符串
本次提交将会话、消息、用量等核心表主键统一为 Int64 雪花ID（DataScale="time"），去除自增，提升分布式扩展性和排序能力。消息反馈表主键调整为 Int32 并保留自增。相关实体类同步调整 Id 类型及查找逻辑，索引和高级查询均切换为基于 Id 字段。前端所有与 id 相关的类型、API、组件参数等由 number 全面切换为 string，保证前后端一致。新增数据清理方法和部分后台管理查询。整体提升系统大数据和多节点场景下的性能与一致性。

+++ NewLife.ChatAI/Entity/消息反馈.cs
@@ -24,13 +24,13 @@ namespace NewLife.ChatAI.Entity;
 public partial class MessageFeedback : IEntity<MessageFeedbackModel>
 {
     #region 属性
-    private Int64 _Id;
+    private Int32 _Id;
     /// <summary>编号</summary>
     [DisplayName("编号")]
     [Description("编号")]
     [DataObjectField(true, true, false, 0)]
     [BindColumn("Id", "编号", "")]
-    public Int64 Id { get => _Id; set { if (OnPropertyChanging("Id", value)) { _Id = value; OnPropertyChanged("Id"); } } }
+    public Int32 Id { get => _Id; set { if (OnPropertyChanging("Id", value)) { _Id = value; OnPropertyChanged("Id"); } } }
 
     private Int64 _MessageId;
     /// <summary>消息。被反馈的消息</summary>
@@ -173,7 +173,7 @@ public partial class MessageFeedback : IEntity<MessageFeedbackModel>
         {
             switch (name)
             {
-                case "Id": _Id = value.ToLong(); break;
+                case "Id": _Id = value.ToInt(); break;
                 case "MessageId": _MessageId = value.ToLong(); break;
                 case "UserId": _UserId = value.ToInt(); break;
                 case "FeedbackType": _FeedbackType = value.ToInt(); break;
@@ -198,7 +198,7 @@ public partial class MessageFeedback : IEntity<MessageFeedbackModel>
     /// <summary>根据编号查找</summary>
     /// <param name="id">编号</param>
     /// <returns>实体对象</returns>
-    public static MessageFeedback FindById(Int64 id)
+    public static MessageFeedback FindById(Int32 id)
     {
         if (id < 0) return null;
